merge-one-file: libify
This moves merge_strategies_one_file() (and its helper functions) to a
new file, merge-strategies.c.  This will enable the resolve and octopus
strategies to directly call it instead of forking.

This is not a faithful copy-and-paste; in the builtin versions,
merge_strategies_one_file() operated on `the_repository' and
`the_index', something we cannot allow a function part of libgit.a to
do.  Hence, it now takes a pointer to a repository as its first
argument (and helper functions takes a pointer to an `index_state').

Signed-off-by: Alban Gruin <alban.gruin@gmail.com>