merge-index: add a new way to invoke `git-merge-one-file'
Since `git-merge-one-file' will be rewritten and libified, there may be
cases where there is no executable named this way (ie. when git is
compiled with `SKIP_DASHED_BUILT_INS' enabled).  This adds a new way to
invoke this particular program even if it does not exist, by passing
`--use=merge-one-file' to merge-index.  For now, it still forks.

The test suite and shell scripts (git-merge-octopus.sh and
git-merge-resolve.sh) are updated to use this new convention.

Signed-off-by: Alban Gruin <alban.gruin@gmail.com>